requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
torch==2.1.0
facenet-pytorch>=2.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
pillow>=9.5.0
faiss-cpu==1.7.4
hnswlib==0.7.1
sqlalchemy==2.2.2
python-multipart==0.0.6
